refactor(ecc): migrate GPU tracegen files into ecc circuit folder
