+++ extensions/algebra/circuit/src/cuda/mod.rs
@@ -19,12 +19,8 @@ use openvm_stark_backend::prover::AirProvingContext;
 use crate::modular_chip::ModularIsEqualCoreCols;
 
 mod cuda_abi;
-/// Exposed for the `EC_MUL` chip, whose projection shape the vec-heap gather cannot describe
-/// and whose crate has no CUDA build of its own, so its kernels are compiled here.
-pub use cuda_abi::{
-    ec_mul_projective_generate_vars, ec_mul_tracegen, gather_ec_mul, merge_range_counts,
-    EcMulFillLaunchConfig,
-};
+/// Exposed for the `EC_MUL` chip's range-checker merge.
+pub use cuda_abi::merge_range_counts;
 pub mod field_expr;
 pub(crate) mod modular_addsub;
 pub(crate) mod vec_heap;
